+++ SQLiteStudio3/guiSQLiteStudio/completer/completerwindow.cpp
@@ -236,9 +236,9 @@ QString CompleterWindow::getStatusMsg(const QModelIndex& index)
         case ExpectedToken::OPERATOR:
             return tr("Operator: %1", "completer statusbar").arg(value);
         case ExpectedToken::STRING:
-            return tr("String", "completer statusbar");
+            return tr("String", "completer statusbar").arg(value);
         case ExpectedToken::NUMBER:
-            return tr("Number", "completer statusbar").arg(value);
+            return tr("Number", "completer statusbar");
         case ExpectedToken::BLOB:
             return tr("Binary data", "completer statusbar").arg(value);
         case ExpectedToken::COLLATION: